Replaced tabs by spaces

+++ moviepy/video/fx/margin.py
@@ -32,7 +32,9 @@ def margin(clip, mar=None, left=0, right=0, top=0,
         new_w, new_h = w + left + right, h + top + bottom
         if clip.ismask:
             shape = (new_h, new_w)
-            bg = np.tile(opacity, (new_h, new_w)).reshape(shape)
+            bg = ( np.tile(opacity, (new_h, new_w))
+                       .astype(float)
+                       .reshape(shape))
         else:
             shape = (new_h, new_w, 3)
             bg = np.tile(color, (new_h, new_w)).reshape(shape)
